ArgoData.jl

Argo data processing and analysis. Currently provides:

The GDAC module functions access and retrieve files from the Argo data servers.

MITprof supports the format of Forget, et al 2015 for standard depth data sets like this one.

AnalysisMethods compute e.g. cost functions and geospatial statistics.

The MITprof_plots module in examples/ is to plot results.

This package is in early developement stage when breaking changes can be expected.

Workflows

Downloading and accessing an Argo file (wmo=13857 from folder="aoml") is done like this.

using ArgoData
input_file=GDAC.download_file("aoml",13857)

using NCDatasets
Dataset(input_file)

A list of all folder,wmo pairs can be obtained using files_list=GDAC.files_list(). And a method to download files in bulk & parallel is presented in examples/Argo_distributed_download.jl.

MITprof Format

Formatting an Argo file (input_file) into an MITprof file (output_file) proceeds as follows.

gridded_fields=GriddedFields.load()
output_file=MITprof.format(gridded_fields,input_file)

The file generated by the previous command can be accessed normally as a NetCDF file (e.g., Dataset(output_file)) or using the convenient MITprofStandard data structure.

mp=MITprofStandard(output_file)

For additional detail, please refer to the examples/ArgoToMITprof.jl example.

Download MITprof files

The original collection of MITprof files from Forget, et al 2015 is archived here. These files can be retrieved as follows.

using CSV, DataFrames
tmp = CSV.File("examples/dataverse_files.csv") |> DataFrame
url0="https://dataverse.harvard.edu/api/access/datafile/"
run(`wget --content-disposition $(url0)$(tmp[1,:ID])`)

Functions

ArgoData.MITprofStandard

ArgoData.ProfileNative

ArgoData.ProfileStandard

ArgoData.AnalysisMethods.cost_functions

ArgoData.AnalysisMethods.profile_add_level!

ArgoData.AnalysisMethods.profile_levels

ArgoData.AnalysisMethods.profile_positions

ArgoData.AnalysisMethods.profile_subset

ArgoData.AnalysisMethods.profile_trim

ArgoData.AnalysisMethods.profile_variables

ArgoData.GDAC.download_file

ArgoData.GDAC.download_file

ArgoData.GDAC.files_list

ArgoData.GDAC.files_list

ArgoData.GDAC.grey_list

ArgoData.GDAC.grey_list

ArgoData.GDAC.wget_geo

ArgoData.MITprof.format

ArgoData.MITprof.format

ArgoData.MITprof.format_loop

ArgoData.MITprof.write

ArgoData.MITprof.write

ArgoData.MITprof.write

ArgoData.ProfileNative — Type

ProfileNative

Container for a multivariate profile read from a GDAC Argo file.

1D arrays: lon,lat,date,ymd,hms,pnumtxt,direc,DATAMODE,isBAD

2D arrays: T,S,pressure,depth,T_ERR,SERR

ArgoData.ProfileStandard — Type

ProfileNative

Container for a multivariate profile in MITprof format.

2D arrays: T,S,Testim,Sestim,Tweight,Sweight,Ttest,Stest,T_ERR,SERR

ArgoData.MITprofStandard — Type

MITprofStandard

Container for a MITprof format file data.

filename : file name

1D arrays: lon,lat,date,depth,ID

2D arrays: T,S,Te,Se,Tw,Sw

ArgoData.GDAC.download_file — Function

download_file(file::DataFrameRow,suff="prof",ftp=missing)

Get folder and wmo from data frame row and them call download_file.

using ArgoData
files_list=GDAC.files_list()
file=GDAC.download_file(files_list[10000,:])

ArgoData.GDAC.download_file — Function

download_file(folder::String,wmo::Int,suff="prof",ftp=missing)

Download Argo file from the GDAC server (<ftp://ftp.ifremer.fr/ifremer/argo/dac/> by default) to a temporary folder (joinpath(tempdir(),"Argo_DAC_files"))

The file name is given by folder and wmo. For example, 13857_prof.nc for wmo=13857 is from the aoml folder.

The default for suff is "prof" which means we'll download the file that contains the profile data. Other possible choices are "meta", "Rtraj", "tech".

If the ftp argument is omitted or isa(ftp,String) then Downloads.download is used. If, alternatively, isa(ftp,FTP) then FTPClient.download is used.

Example :

using ArgoData
GDAC.download_file("aoml",13857)

#or:
ftp="ftp://usgodae.org/pub/outgoing/argo/dac/"
GDAC.download_file("aoml",13857,"meta",ftp)

ArgoData.GDAC.files_list — Method

files_list(fil::String)

Get list of Argo float files from csv file with columns two columns – folder and wmo.

using ArgoData
fil="https://raw.githubusercontent.com/JuliaOcean/ArgoData.jl/gh-pages/dev/Argo_float_files.csv"
files_list=GDAC.files_list(fil)

ArgoData.GDAC.files_list — Method

files_list()

Get list of Argo float files from Ifremer GDAC server ftp://ftp.ifremer.fr/ifremer/argo/dac/

using ArgoData
files_list=GDAC.files_list()

ArgoData.GDAC.grey_list — Method

grey_list(fil::String)

Read "ar_greylist.txt" file into a DataFrame.

ArgoData.GDAC.grey_list — Method

grey_list()

Download "ar_greylist.txt" from GDAC and read file into a DataFrame.

ArgoData.GDAC.wget_geo — Method

wget_geo(b::String,y::Int,m::Int)

Download, using wget, Argo data files for one regional domain (b), year (y), and month (m) from the GDAC FTP server (ftp://ftp.ifremer.fr/ifremer/argo or, equivalently, ftp://usgodae.org/pub/outgoing/argo).

b="atlantic"; yy=2009:2009; mm=8:12;
for y=yy, m=mm;
    println("$b/$y/$m"); GDAC.wget_geo(b,y,m)
end

ArgoData.MITprof.format — Function

format(meta,gridded_fields,input_file,output_file="")

From Argo file name as input : read input file content, process into the MITprof format, and write to MITprof file.

MITprof.format(meta,gridded_fields,input_file)

ArgoData.MITprof.format — Method

format(gridded_fields,input_file)

From Argo file name as input : read input file content, process into the MITprof format, and write to MITprof file.

MITprof.format(gridded_fields,input_file)

ArgoData.MITprof.format_loop — Method

format_loop(II)

Loop over files and call format.

gridded_fields=GriddedFields.load()
fil=joinpath(tempdir(),"Argo_MITprof_files","input","Argo_float_files.csv")
files_list=GDAC.files_list(fil)
MITprof.format_loop(gridded_fields,files_list,1:10)

ArgoData.MITprof.write — Method

MITprof.write(meta,profiles,profiles_std;path="")

Create an MITprof file from meta data + profiles during MITprof.format.

MITprof.write(meta,profiles,profiles_std)

ArgoData.MITprof.write — Method

write(fil::String,mp::MITprofStandard)

Create an MITprof file from an MITprofStandard input.

ArgoData.MITprof.write — Method

write(fil::String,mps::Vector{MITprofStandard})

Create an MITprof file from a vector of MITprofStandard inputs.

ArgoData.AnalysisMethods.cost_functions — Function

cost_functions(vv="prof_T",JJ=[])

Loop through files and compute nb profiles, nb non-blank profiles, nb levels mean, cost mean.

pth="MITprof/"
nt,np,nz,cost=MITprof.cost_functions(pth,"prof_S")

using JLD2
jldsave(joinpath("csv","prof_S_stats.jld2"); nt,np,nz,cost)

ArgoData.AnalysisMethods.profile_add_level! — Method

profile_add_level!(df,k)

df=CSV.read("csv/profile_positions.csv",DataFrame)
MITprof.profile_add_level!(df,5)

ArgoData.AnalysisMethods.profile_levels — Function

profile_levels()

Create Array of all values for one level, obtained by looping through files in csv/.

ArgoData.AnalysisMethods.profile_positions — Method

profile_positions(path)

Create table (DataFrame) of the positions and dates obtained by looping through files in path. Additional information such as float ID, position on the ECCO grid pos, number of valid data points for T and S (nbT ,nbS).

using ArgoData
path="MITprof/"
csv_file="csv/profile_positions.csv"

using MeshArrays
γ=GridSpec("LatLonCap",MeshArrays.GRID_LLC90)
Γ=GridLoad(γ)

df=MITprof.profile_positions(path,Γ)
CSV.write(csv_file, df)

ArgoData.AnalysisMethods.profile_subset — Method

profile_subset(df,lons,lats,dates)

df=CSV.read("csv/profile_positions.csv",DataFrame)
d0=DateTime("2012-06-11T18:50:04")
d1=DateTime("2012-07-11T18:50:04")
tmp=MITprof.profile_subset(df,(0,10),(-5,5),(d0,d1))

ArgoData.AnalysisMethods.profile_trim — Method

profile_trim(df)

ArgoData.AnalysisMethods.profile_variables — Method

profile_variables(name::String)

Create Array of all values for one variable, obtained by looping through files in path.

@everywhere using ArgoData, CSV, DataFrames
@everywhere list_v=("prof_T","prof_Testim","prof_Tweight","prof_S","prof_Sestim","prof_Sweight")
@distributed for v in list_v
    output_file="csv/"*v*".csv"
    tmp=MITprof.profile_variables(v)
    CSV.write(output_file,DataFrame(tmp,:auto))
end