ArgoData.jl

This package is at a very early stage of development. Stay tuned ...

ArgoData.DownloadArgo.GDAC_FTP

ArgoData.DownloadArgo.GetOneProfile

ArgoData.DownloadArgo.mitprof_interp_setup

ArgoData.DownloadArgo.prof_PtoZ!

ArgoData.DownloadArgo.prof_TtoΘ!

ArgoData.DownloadArgo.prof_interp!

ArgoData.DownloadArgo.sw_adtg

ArgoData.DownloadArgo.sw_dpth

ArgoData.DownloadArgo.sw_ptmp

ArgoData.MITprof.loop

ArgoData.MITprof.read

ArgoData.MITprof.loop — Function

loop(pth::String="profiles/")

Standard Depth Argo Data Collection – see ?MITprof.read for detail.

ArgoData.MITprof.read — Function

read(f::String="MITprof/MITprof_mar2016_argo9506.nc")

Standard Depth Argo Data Example.

Here we read the MITprof standard depth data set from https://doi.org/10.7910/DVN/EE3C40 For more information, please refer to Forget, et al 2015 (http://dx.doi.org/10.5194/gmd-8-3071-2015) The produced figure shows the number of profiles as function of time for a chosen file and maps out the locations of Argo profiles collected for a chosen year.

using ArgoData, Plots

fi="MITprof/MITprof_mar2016_argo9506.nc"
(lo,la,ye)=MITprof.read(fi)

h = histogram(ye,bins=20,label=fi[end-10:end],title="Argo profiles")

ye0=2004; ye1=ye0.+1
kk=findall((ye.>ye0) .* (ye.<ye1))
scatter(lo[kk],la[kk],label=fi[end-10:end],title="Argo profiles count")

ArgoData.DownloadArgo.GDAC_FTP — Method

GDAC_FTP(b::String,y::Int,m::Int)

Download Argo data files for one regional domain (b), year (y), and month (m) from the GDAC FTP server (ftp://ftp.ifremer.fr/ifremer/argo or, equivalently, ftp://usgodae.org/pub/outgoing/argo).

b="atlantic"; yy=2009:2009; mm=8:12;
for y=yy, m=mm;
    println("$b/$y/$m"); DownloadArgo.GDAC_FTP(b,y,m)
end

ArgoData.DownloadArgo.GetOneProfile — Method

GetOneProfile(m)

Get one profile from netcdf file.

ArgoData.DownloadArgo.mitprof_interp_setup — Method

mitprof_interp_setup(fil::String)

Get parameters etc from yaml file (fil).

ArgoData.DownloadArgo.prof_PtoZ! — Method

prof_PtoZ!(prof,meta)

Convert prof["p"] to prof["depth"]

ArgoData.DownloadArgo.prof_TtoΘ! — Method

prof_TtoΘ!(prof,meta)

Convert prof["T"] to potential temperature

ArgoData.DownloadArgo.prof_interp! — Method

prof_interp!(prof,meta)

Interpolate from prof["depth"] to meta["z_std"]

ArgoData.DownloadArgo.sw_adtg — Method

sw_adtg(S,T,P)

Calculate adiabatic temperature gradient as per UNESCO 1983 routines from salinity (S; in psu), in situ temperature (T; in °C), and pressure (P; in decibar)

adtg = sw_adtg(S,T,P)

ArgoData.DownloadArgo.sw_dpth — Method

sw_dpth(P,LAT)

Calculate depth in meters from pressure (P; in decibars) and latitude (LAT; in °N)

d = sw_dpth(100.0,20.0)

ArgoData.DownloadArgo.sw_ptmp — Function

sw_ptmp(S,T,P,PR)

Calculate potential temperature as per UNESCO 1983 report from salinity (S; in psu), in situ temperature (T; in °C), and pressure (P; in decibar) relative to PR (in decibar; 0 by default).

ptmp = sw_ptmp(S,T,P,PR=missing)