ArgoData.jl

Argo data processing and analysis.

The GDAC module functions access and retrieve files from the Argo data servers.

MITprof supports the format of Forget, et al 2015 for standard depth data sets like this one.

This package is in early developement stage when breaking changes can be expected.

ArgoData.GDAC.Argo_float_download

ArgoData.GDAC.Argo_float_files

ArgoData.GDAC.Argo_float_files

ArgoData.GDAC.greylist

ArgoData.GDAC.greylist

ArgoData.GDAC.wget_geo

ArgoData.MITprof.MITprof_format

ArgoData.MITprof.MITprof_read

ArgoData.MITprof.MITprof_write

ArgoData.MITprof.loop

Workflows

Downloading and accessing an Argo file can simply be done like this.

using Downloads, NCDatasets

wmo=6900900
url0="https://data-argo.ifremer.fr/dac/coriolis/"
input_url=url0*"/$(wmo)/$(wmo)_prof.nc"
input_file=joinpath(tempdir(),"$(wmo)_prof.nc")
fil=Downloads.download(input_url,input_file)
Dataset(fil)

Formatting of an Argo file (input_file) into an MITprof file (output_file) proceeds as follows.

using ArgoData

meta=ArgoTools.meta(input_file,output_file)
gridded_fields=GriddedFields.load()
MITprof.MITprof_format(meta,gridded_fields,input_file,output_file)

For additional detail, please refer to the examples/ArgoToMITprof.jl example.

Functions

ArgoData.MITprof.MITprof_format — Function

MITprof_format(meta,gridded_fields,input_file,output_file="")

From Argo file name as input : read input file content, process into the MITprof format, and write to MITprof file.

MITprof.MITprof_format(meta,gridded_fields,input_file)

ArgoData.MITprof.MITprof_read — Function

MITprof_read(f::String="MITprof/MITprof_mar2016_argo9506.nc")

Standard Depth Argo Data Example.

Here we read the MITprof standard depth data set from https://doi.org/10.7910/DVN/EE3C40 For more information, please refer to Forget, et al 2015 (http://dx.doi.org/10.5194/gmd-8-3071-2015) The produced figure shows the number of profiles as function of time for a chosen file and maps out the locations of Argo profiles collected for a chosen year.

using ArgoData, Plots

fi="MITprof/MITprof_mar2016_argo9506.nc"
(lo,la,ye)=MITprof.MITprof_read(fi)

h = histogram(ye,bins=20,label=fi[end-10:end],title="Argo profiles")

ye0=2004; ye1=ye0.+1
kk=findall((ye.>ye0) .* (ye.<ye1))
scatter(lo[kk],la[kk],label=fi[end-10:end],title="Argo profiles count")

ArgoData.MITprof.MITprof_write — Method

MITprof.MITprof_write(meta,profiles,profiles_std;path="")

Write to file.

MITprof.MITprof_write(meta,profiles,profiles_std)

ArgoData.MITprof.loop — Function

loop(pth::String="profiles/")

Standard Depth Argo Data Collection – see ?MITprof.read for detail.

ArgoData.GDAC.Argo_float_download — Function

Argo_float_download(list_files,ii=1,suff="prof",ftp=missing)

Download one Argo file for float ranked ii in list_files from GDAC server (ftp://ftp.ifremer.fr/ifremer/argo/dac/ by default) to a temporary folder (joinpath(tempdir(),"Argo_DAC_files")). By default suff="prof" means we'll download the file that contains the profile data (e.g. 13857_prof.nc for ii=1 with wmo=13857). Other possible choices for suff: "meta", "Rtraj", "tech". If the ftp argument is omitted or isa(ftp,String) then Downloads.download is used. If, alternatively, isa(ftp,FTP) then FTPClient.download is used.

Example :

using ArgoData
list_files=GDAC.Argo_float_files()
GDAC.Argo_float_download(list_files,10000)

ftp="ftp://usgodae.org/pub/outgoing/argo/dac/"
GDAC.Argo_float_download(list_files,10000,"meta",ftp)

ArgoData.GDAC.Argo_float_files — Method

Argo_float_files(fil::String)

Get list of Argo float files from csv file with columns two columns – folder and wmo.

using ArgoData
fil="https://gaelforget.github.io/OceanRobots.jl/dev/examples/Argo_float_files.csv"
GDAC.Argo_float_files(fil)

ArgoData.GDAC.Argo_float_files — Method

Argo_float_files()

Get list of Argo float files from Ifremer GDAC server ftp://ftp.ifremer.fr/ifremer/argo/dac/

using ArgoData
GDAC.Argo_float_files()

ArgoData.GDAC.greylist — Method

greylist(fil::String)

Read "ar_greylist.txt" file into a DataFrame.

ArgoData.GDAC.greylist — Method

greylist()

Download "ar_greylist.txt" from GDAC and read file into a DataFrame.

ArgoData.GDAC.wget_geo — Method

wget_geo(b::String,y::Int,m::Int)

Download, using wget, Argo data files for one regional domain (b), year (y), and month (m) from the GDAC FTP server (ftp://ftp.ifremer.fr/ifremer/argo or, equivalently, ftp://usgodae.org/pub/outgoing/argo).

b="atlantic"; yy=2009:2009; mm=8:12;
for y=yy, m=mm;
    println("$b/$y/$m"); GDAC.wget_geo(b,y,m)
end

Missing docstring.

Missing docstring for ProfileNative. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ProfileStandard. Check Documenter's build log for details.